Refactor: Create dropdown menu component
