+++ src/components/DropdownMenu/DropdownMenu.module.css
@@ -1,0 +1,45 @@
+.dropdownTrigger {
+  color: var(--ds-color-main-text-default);
+  margin-top: 30px;
+  background-color: transparent;
+  border: none;
+}
+
+.dropdownTrigger:hover,
+.dropdownTrigger:focus,
+.dropdownTrigger:active,
+.dropdownTrigger:focus-visible {
+  background-color: var(--ds-color-neutral-surface-hover);
+  border: none;
+  outline: none;
+}
+
+.dropdownButton {
+  border: none;
+  outline: none;
+}
+
+.dropdownButton:hover {
+  border: none;
+  background-color: var(--ds-color-neutral-surface-hover);
+}
+
+.dropdownLink {
+  text-decoration: none;
+  color: var(--ds-color-main-text-default);
+  padding: 0;
+}
+
+@media (max-width: 768px) {
+  .dropdownTrigger {
+    margin-left: 20px;
+    margin-top: 0;
+    padding-left: 0.2rem;
+  }
+}
+
+@media (max-width: 480px) {
+  .dropdownTrigger {
+    font-size: 0.9rem;
+  }
+}
